removes interfases from service layer
