+++ phonebook-api/src/main/java/com/telran/phonebookapi/service/ConfirmationTokenService.java
@@ -1,13 +1,26 @@
 package com.telran.phonebookapi.service;
 
 import com.telran.phonebookapi.entity.ConfirmationToken;
+import com.telran.phonebookapi.exception.TokenNotFoundException;
+import com.telran.phonebookapi.repository.ConfirmationTokenRepository;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
 
-import java.util.Optional;
+@Service
+@AllArgsConstructor
+public class ConfirmationTokenService {
 
-public interface ConfirmationTokenService {
-    void save(ConfirmationToken confirmationToken);
+    private final ConfirmationTokenRepository confirmationTokenRepository;
 
-    void delete(Long id);
+    public void save(ConfirmationToken confirmationToken) {
+        confirmationTokenRepository.save(confirmationToken);
+    }
 
-    Optional<ConfirmationToken> findByToken(String token);
+    public void delete(Long id){
+        confirmationTokenRepository.deleteById(id);
+    }
+
+    public ConfirmationToken findByToken(String token) throws TokenNotFoundException {
+        return confirmationTokenRepository.findByToken(token).orElseThrow(() -> new TokenNotFoundException("Token not found"));
+    }
 }
