--- phonebook-api/src/main/java/com/telran/phonebookapi/service/EmailSenderService.java
@@ -1,7 +1,0 @@
-package com.telran.phonebookapi.service;
-
-import org.springframework.mail.SimpleMailMessage;
-
-public interface EmailSenderService {
-    void sendEmail(SimpleMailMessage email);
-}
