--- phonebook-api/src/main/java/com/telran/phonebookapi/service/impl/EmailSenderServiceImpl.java
@@ -1,22 +1,0 @@
-package com.telran.phonebookapi.service.impl;
-
-import com.telran.phonebookapi.service.EmailSenderService;
-import lombok.AllArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class EmailSenderServiceImpl implements EmailSenderService {
-
-    private final JavaMailSender javaMailSender;
-
-    @Async
-    @Override
-    public void sendEmail(SimpleMailMessage email) {
-        javaMailSender.send(email);
-    }
-}
-
