+++ phonebook-api/src/main/resources/application.properties
@@ -1,12 +1,12 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/
-spring.datasource.username=
-spring.datasource.password=
+spring.datasource.url=jdbc:postgresql://localhost:5432/person
+spring.datasource.username=postgres
+spring.datasource.password=root
 spring.jpa.hibernate.ddl-auto=create
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=<EMAIL>
-spring.mail.password=<PASSWORD>
+spring.mail.username=<mail>
+spring.mail.password=<password>
 
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
