+++ phonebook-api/src/test/java/com/telran/phonebookapi/service/UserServiceTest.java
@@ -2,13 +2,13 @@ package com.telran.phonebookapi.service;
 
 import com.telran.phonebookapi.entity.ConfirmationToken;
 import com.telran.phonebookapi.entity.User;
+import com.telran.phonebookapi.exception.TokenNotFoundException;
 import com.telran.phonebookapi.repository.UserRepository;
-import com.telran.phonebookapi.service.impl.UserServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -25,10 +25,7 @@ class UserServiceTest {
     @Mock
     private BCryptPasswordEncoder bCryptPasswordEncoder;
     @Mock
-    private EmailSenderService emailSenderService;
-
-    @InjectMocks
-    UserServiceImpl userService;
+    private JavaMailSender javaMailSender;
 
     @Test
     public void testCreate_newUser_returnsOk() {
@@ -42,13 +39,15 @@ class UserServiceTest {
     }
 
     @Test
-    public void testConfirm_validToken_returnsOk() {
+    public void testConfirm_validToken_returnsOk() throws TokenNotFoundException {
         User user = new User("anna@gmail.com", "kjkjsdfsdfdf");
         user.setEnabled(true);
         ConfirmationToken token = new ConfirmationToken(user);
+        confirmationTokenService.findByToken(token.getToken());
         userRepository.save(user);
         confirmationTokenService.delete(token.getId());
 
+        verify(confirmationTokenService, times(1)).findByToken(any());
         verify(userRepository, times(1)).save(any());
         verify(confirmationTokenService, times(1)).delete(any());
     }
